+++ app/validators/register.ts
